+++ packages/core/src/extensions/extension.test.ts
@@ -142,6 +142,8 @@ describe('defineEditorExtension', () => {
           },
           {
             "attrs": {
+              "fenceLength": null,
+              "fenceStyle": null,
               "language": "ts",
             },
             "content": [